Bug 1713715. Redirect double tap gestures to the root content apzc. r=botond
This means that double tapping on an oop iframe "works" in that we don't try (and fail) to zoom the oop iframe. But it also means it is impossible to zoom to some content inside that oop iframe. Instead the best we can do is zoom to the oop iframe itself. This is consistent with what Chrome and Safari currently do. Allowing zooming to content inside an oop iframe would be decently complicated and it doesn't seem worth it. Bug 1715179 is on file for this.

Differential Revision: https://phabricator.services.mozilla.com/D116925

UltraBlame original commit: 5785ea7b467a83ff9d1fee70f93dff1bf196ddb2