Bug 1715512 part 15 - Fix tests. r=jonco
The jit-test was triggering slices that were "too long" for the test to abort the
sweep phase. Reduce the count for the test to pass.

The devtools test was failing because shapes no longer show up in the visible part
of the memory allocation table (it's more down the list because there are fewer
shapes). Use PropMap instead.

Differential Revision: https://phabricator.services.mozilla.com/D117315

UltraBlame original commit: 5b1e0b26ed64c88160315e6f70e818d620777ae3