Bug 1715783 - Use nsCSSRendering::DetermineBackgroundColor in backplate color computation. r=morgan
This is more correct as it takes care about falling back to white as needed in
printing mode.

This is still not perfect because themed frames still get black-on-black text.

I'm not super-sure what's the right way to fix that, will
file a follow-up for that. In particular, when you have a themed
button, we use the system colors (i.e., black) to paint
it, but nsLayoutUtils::DarkenColorIfNeeded doesn't know it
and darkens the text color anyways.

Easiest fix is just not use the system colors when
printing, but that feels not-amazing...

Differential Revision: https://phabricator.services.mozilla.com/D118275

UltraBlame original commit: b4e4f16482fe83abebdf70ca4f947c331b33244d