Bug 1647851 Part 3 - Add basic tests for applying page-size when printing r=dholbert
These aren't WPT because the CSS-page spec gives a large range of correct
behavior for handling page-size when printing. We also don't expect Firefox's
default handling to remain the same as it is implemented in this bug.

These also can't be paged reftests because of bug 1717138

Ideally we would be able to give a range of possible correct results for a WPT
that tests printing, but for now we are just testing the specific behavior we
implement.

Differential Revision: https://phabricator.services.mozilla.com/D117480

UltraBlame original commit: 4aee63294f3940bf9b80611a4e6989700a558386