Bug 1717584 - [devtools] Do not reuse destroyed frame actors in Thread actor r=ochameau
Depends on D118470

The ThreadActor is storing frame actors in a local map, on top of the regular pools managed by the framework.
If a frame corresponding to an already destroyed actor needs to be handled and returned again by the thread actor, it should create a completely new Actor instead of attempting to reuse an already destroyed object.

Differential Revision: https://phabricator.services.mozilla.com/D118471

UltraBlame original commit: ff8225deba22d2d84d8678851a5589b75845634c