Bug 1717414 - [profiler] Add a new property to network markers to report the new channel id for redirection processes r=gerald
Then this will make it possible in the UI to relate the redirected
request with its redirection.

With this patch a debug build of Firefox will crash when a redirection
happens, but the next patch will fix this.

Differential Revision: https://phabricator.services.mozilla.com/D118467

UltraBlame original commit: 0083df7a6f653cea1e415a023be4b681c1baf834