Bug 1716712 - Use JIT information in DoFramePointerBacktrace on x64 - r=canaltinova
This is similar to what DoMozStackWalkBacktrace does, by looking where the stack walker stopped, and if it can be restarted at a calling C++-to-JIT entry frame.

Note: The code is pretty much a modified copy, but factoring out the common parts would be difficult because of the specific local variables, the different stack-walking functions, and the different register storages.

Also, instead of mentioning the OS (was only Windows), the choice of platform support is now based on the actual Profiler stack walkers that are used.

Differential Revision: https://phabricator.services.mozilla.com/D118109

UltraBlame original commit: f12d9f8337b0d3184aa69bc9c4f60ac1cd806517