Bug 1715970 [wpt PR 29342] - Always use the referrer used to load the document on same-document navigations, a=testonly
Automatic update from web-platform-tests
Always use the referrer used to load the document on same-document navigations

DidCommitProvisionalLoadParams has the `referrer` param that is used
to save the `referrer` attribute in FrameNavigationEntry after commit.

Currently, most same-document navigations will use the initial referrer
that's used to load the document, even after several navigations. An
exception to this is if the navigation is classified as a client-side
redirect, in which case the previous URL is used. This seems to be
accidental, as the `referrer_` that's tracked in the DocumentLoader is
never actually updated for these navigations, so future same-document
navigations will still use the initial referrer that's used to load the
document.

This CL removes the special behavior for client-side redirect same-doc
navigations, to make the behavior more predictable, making it easier
to calculate the `referrer` param in the browser in the future, and also
allows us to remove the redirect chain in the renderer completely in
crrev.com/c/2933038.

Bug: 1171210, 1131832
Change-Id: If460d5fe08ab5795b3dd4d69b0188289d8253839
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2938762
Commit-Queue: Rakina Zata Amni <rakinachromium.org>
Reviewed-by: David Van Cleve <davidvcchromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzognichromium.org>
Reviewed-by: Antonio Sartori <antoniosartorichromium.org>
Cr-Commit-Position: refs/heads/master{#892950}

--

wpt-commits: 50b6ca1cfe9230c462032b168e8eabaf3626497e
wpt-pr: 29342

UltraBlame original commit: af1cb445717d8e42c2517a20ad3f532462118fd6