Bug 1716013 [wpt PR 29345] - [composite-bgcolor-animation] Return empty result when PaintWorkletInput not found, a=testonly
Automatic update from web-platform-tests
[composite-bgcolor-animation] Return empty result when PaintWorkletInput not found

The root cause of the problem is that there is an early exit in the
DiscardableImageMap::GatherDiscardableImages(), happening when the
paint rect for the op is empty. This makes sense because there is
nothing to draw.

Note that this is different than the empty element case. In fact,
we already have "background-color-animation-zero-size-element.html"
under css-backgrounds/animations/ and there is no problem with that.
The bug happens when the element is not empty, but its intersection
with the SkCanvas that it draws into is empty. Please see the notes
in this doc for more details:
https://docs.google.com/document/d/1Da9d7ai5xt4nPX7NDZnqaaTSoFitRSeYxw1qxGjtdeQ/edit

The |PaintWorkletImageProvider::records_| contains the records for
all the discardable images that can draw. Meaning that the ones that
early exits won't be included in the |records_|. But, when the
PaintWorkletImageProvider::GetPaintRecordResult() is called, the
input parameter |PaintWorkletInput| would contain all possible
PaintWorkletInput that is passed from Blink to CC, which obviously
contains the ones that cannot draw. As a result, these
PaintWorkletInputs won't exists in the |records_|.

We handle these cases by returning an empty record because there
is nothing to draw (like empty container cases).

Bug: 1216210
Change-Id: I9630aa4b55fd1322f641f42b05c20aab8c7234c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2954479
Reviewed-by: Philip Rogers <pdrchromium.org>
Reviewed-by: Robert Flack <flackrchromium.org>
Commit-Queue: Xida Chen <xidachenchromium.org>
Cr-Commit-Position: refs/heads/master{#894050}

--

wpt-commits: 4c83511984d28d692a68de2330a844183b9d488e
wpt-pr: 29345

UltraBlame original commit: 136c0fe36862d59b4bcd953ef538c1fc5fcdc813