Bug 1697639 - Only add AGRs to mFrameToAnimatedGeometryRootMap when creating AGRs r=mattwoodrow
This enforces a stable frame to AGR mapping, which was previously changing when AGR status of the frame changed due to display port change.
The downside of this is that the intermediate results are not cached, which means more traversals.

Differential Revision: https://phabricator.services.mozilla.com/D118603

UltraBlame original commit: eb9c0b2c751632402eceaa47b60f5a2fddb8e3cf