Bug 1718102 - Extract ScriptFlags accessors to be reused. r=arai
The CompilationInput currently let the consumer reach to the lazy field, to
reach accessors implemented on the BaseScript. As we are trying to remove direct
usage of the lazy field, and later abstract between a ScriptStencilExtra or a
BaseScript, we have to make it simple to add these accessors to other
structures.

This change extract the accessors and move them to a parametrized structure
which knows that its parameters is able to locate the field which contains all
these flags.

To implement accessors, one should inherit from the class
  ImmutableScriptFlagsAccessors<...>

and parametrize it with the derived class as long as the derived class
implements the function

  ImmutableFlags immutableFlags() const { ... }

And similarly for mutable flags, except that the returned value should be a
reference.

Differential Revision: https://phabricator.services.mozilla.com/D119105

UltraBlame original commit: 55372527ebed955e8ae108b0e3592ff8b34f967c