Bug 1718934 - Parse -webkit-focus-ring-color (computing to black). r=mstange,geckoview-reviewers,aklotz
Some pages use this to hide the focus outline. On Blink it seems black,
WebKit uses it to expose the OS accent color. Make it black, which is
the default when the color isn't implemented on a given platform.

Differential Revision: https://phabricator.services.mozilla.com/D119036

UltraBlame original commit: d06a9296cfab476b49da4fda4e8b720bf0bf826c