Bug 1692749 - Skip superfluous about:blank load when duplicating a tab, r=kashav,extension-reviewers,kmag
When we duplicate a tab, we don't need to have about:blank load in it, because
we are going to use restore mechanism to copy data into the new tab. If we
don't skip the superfluous load, the restoring process might race with the
loading of about:blank, and sometimes we might try to destroy the
WindowGlobalChild actor just as SessionStore is trying to restore docshell
capabilities for that tab resulting in a rejected promise in _restoreHistory
and `_restoreHistoryComplete` not getting called.

Differential Revision: https://phabricator.services.mozilla.com/D119313

UltraBlame original commit: f3dd29c3cf52c7e1cf268629306b7d1943c54898