Bug 1706268 - Add telemetry probes to collect usage of `Document.execCommand` etc with Gecko-specific commands r=smaug, data-review=travis_
Currently, `increasefontsize`, `decreasefontsize`, `gethtml`, `heading`,
`contentReadOnly`, `readonly` and `insertBrOnReturn` commands are supported
only by Gecko.  So, if nobody uses some or all of them, we can drop the
unused commands.

For saving the space of `Document.mUseCounters`, this patch groups
`Document.queryCommandState` and `Document.queryCommandValue` because they
are both getter of the command.  The difference is not important.  And also
this patch groups `Document.queryCommandSupported` and
`Document.queryCommandEnabled` because they may be used for feature detection
and the difference is not important.

Differential Revision: https://phabricator.services.mozilla.com/D118956

UltraBlame original commit: dbe0910492451298e401178562a336899f8618e7