Bug 1638097 - Add a temporarilyInstalled boolean property to WebExtensionPolicy. r=mixedpuppy
This patch adds a new `temporarilyInstalled` property to the WebExtensionPolicy object, set to
the value of the same property from a Extension.jsm perspective.

The new property is being introduced mainly to be used in a patch that follows to allow the
ServiceWorkerManager internals to access that from the C++ layer (but it may be also
become a useful knowledge for other C++ components in the future).

Differential Revision: https://phabricator.services.mozilla.com/D119529

UltraBlame original commit: c5a37c27cf88bd3c65dc1e155d6bdaaf24bd6194