Bug 1694496 - Annotate FOG JS PHF entries for their widths r=janerik
Without these annotations the literals will be assumed to be `int` so as soon
as there's a literal that cannot be represented as such it'll warn.
And CI builds with warnings as errors.

`metric_entry_t` is `uint64_t` which is `unsigned long long int` which is `ull`
`category_entry_t` is `uint32_t` which is `unsigned long int` which is `ul`

Differential Revision: https://phabricator.services.mozilla.com/D119978

UltraBlame original commit: 628fd187bba918fbd87e5c7e24a27dd0bae434e0