Bug 1718271 - wasm: Memoize rtt.sub results. r=lth
This commit implements the following section from MVP.md for rtt.sub:
`multiple invocations of this instruction with the same operand yield the same observable RTTs`
The typing of the rtt.sub instruction was also outdated and is updated.

There are two inputs to rtt.sub. An immediate type index that the output
'sub-rtt' should have, and an operand rtt value that that is the 'parent-rtt'.

RttValue is updated to now contain both a 'Parent' strong pointer,
and a 'Children' weak hashmap indexed by the canonical rtt of a
type index.

rtt.sub now first gets the canonical rtt of the type index of the
output 'sub-rtt'. Then it looks up if the parent has already generated
a sub-rtt for this type index, and returns it if so. Otherwise a
new sub-rtt is created and inserted into the weakmap.

This is the same approach that V8 has used for this operation,
with the exception that they use a linear search on a strong
vector of child rtt's.

Differential Revision: https://phabricator.services.mozilla.com/D118843

UltraBlame original commit: 3200d5a1f5ae4bc968748fcc730ec944b8440efe