Bug 1717703 Part 1 - Set display list clipping rect for nsPageFrame children to use the content size. r=emilio,mattwoodrow
Without explicitly setting the clipping rect to the specified page size, the
building rect of the display lists will be kept to the size of the nsPageFrame.
This means that any content in the nsPageContentFrame which is larger than the
physical paper size will be clipped.

The actual content frames are scaled down when this is too large, so no
overdraw will occur.

Differential Revision: https://phabricator.services.mozilla.com/D119459

UltraBlame original commit: 1a2bbe2775867a382f6c155c4a50e831e7cd039e