Bug 1717914 - wasm: Rename MaxMemory32* to MaxMemory*. r=lth
The MaxMemory(Pages,Bytes,BoundsCheckLimit) methods are
named to be only applicable for memory32. Currently they are
used for both memory32 and memory64 and I would expect that
to stay that way for a while. So this commit renames them to
be general purpose.

Differential Revision: https://phabricator.services.mozilla.com/D118648

UltraBlame original commit: 4554af03e7c8ff4f642d131e15d1b7c76c9f6d18