Bug 1709640 - Add GeckoView Lite build variant. r=jmaher,aklotz,mhentges,glandium
This patch adds a new GeckoView build variant dubbed "lite". We are in the
process of adding some optional dependencies to GeckoView that are needed by
Fenix (like Glean and Nimbus) which are likely not going to be used by anyone
else at Mozilla. To avoid bloating third party consumers, we provide a
geckoview-lite build that doesn't contain any optional dependency.

The min SDK version of GeckoView (non-lite) is being increased to 21 to match
Glean's min SDK. Given that Fenix's min SDK version is already 21 and that the
lite version's min SDK is still at 16 this change is safe to do and will not
incur in any loss of usability.

Differential Revision: https://phabricator.services.mozilla.com/D114370

UltraBlame original commit: f918535301cbb21e9dfa4ef26eb7e88e632c00d2