Bug 1717625 - [devtools] Fix storage inspector tree when bfCacheInParent is enabled and server side target switching is disabled. r=ochameau.
In this patch, we pass a property alongside the DevToolsFrameChild:bf-cache-navigation-pageshow
message that indicates if a new target is being created as a result of this
bfcache navigation.
This flag is then used in the ParentProcessStorage to spawn a new StorageActorMock
when a new target is being created.

As a result we can stop disabling the bfCacheInParent pref in browser_storage_cookies_navigation.js,
where a bfcache navigation is performed.

Differential Revision: https://phabricator.services.mozilla.com/D120438

UltraBlame original commit: d0fbda14ef14f32303ef184570962c0fa0e33231