Bug 1714639 - Don't call FrameLoader.printPreview in hidden frames. r=mstriemer,jwatt
We need to be displayed when cloning into a same-process frame, so that
the preview code hooks up to the view tree correctly and the frame is
correctly painted.

This works for out-of-process frames, but mostly by chance.

Fortify the printPreview() API code to gracefully handle that case, and
remove the sessionrestore special-case which is not really special.

Differential Revision: https://phabricator.services.mozilla.com/D119980

UltraBlame original commit: f032f4d08cfddac02ef445536ebf259035e498bb