Bug 1671527 Part 1 - Add MaybePushFloatContainingBlock() method. r=emilio
Extract a helper to consolidate the if-else-if logic dealing with float
containing block, and adapt the callers to use it rather than
PushFloatContainingBlock().

This patch shouldn't change the behavior.

Note: In ConstructTableCell(), `!isBlock` is equivalent to
ShouldSuppressFloatingOfDescendants(cellInnerFrame) since
nsMathMLmtdInnerFrame is of eMathML type, so it's OK to just call
MaybePushFloatContainingBlock().

Differential Revision: https://phabricator.services.mozilla.com/D120103

UltraBlame original commit: f2f7f48c0adbaf16a58367c31b773757e36acd69