Bug 1671527 Part 3 - Reparent floats when creating non-column-span wrapper frames. r=emilio
When we are in CreateColumnSpanSiblings(), there are floats in
nsFrameConstructorState::mFloatedList that are waiting to be added to
aInitialBlock once we leave scope of the float containing block.
However, since we are creating non-column-span continuations, which are
continuations of aInitialBlock, we should reparent the floats to these
continuations if their placeholders are the descendants of the
newly-created continuations.

multicol-span-float-002.html exercises
nsCSSFrameConstructor::ConstructBlock() and multicol-span-float-003.html
exercises nsCSSFrameConstructor::CreateIBSiblings().

Without this patch, multicol-span-float-002.html and
multicol-span-float-003.html still pass, but they trigger

```
ASSERTION: nsBlockFrame::CheckFloats: Explicit float list is out of
sync with float cache
```

1524382.html used to trigger the above assertions. Now it's fixed.

Differential Revision: https://phabricator.services.mozilla.com/D120105

UltraBlame original commit: fa7249945de1f8adc586889a4c7fb1f89322a219