Bug 1721700 - Move browser.tabs.unloadOnLowMemory to TabUnloader.unloadTabAsync.  r=haik
We have the pref `browser.tabs.unloadOnLowMemory` to turn on the tab unloading
feature.  Since it controlled whether we initialize the memory watcher or not
in `TabUnloader.init()`, changing it required process start.

To conduct A/B testing for this feature, our infrastructure needs to be able
to change the pref in the middle of session.  So this change moves the pref
check from `TabUnloader.init()` to `TabUnloader.unloadTabAsync()`.  With this,
we fully exercise the memoruy watcher regardless of the pref's value, but invokes
tab unloading only when the pref is on.

Differential Revision: https://phabricator.services.mozilla.com/D120744

UltraBlame original commit: a8d9e102189a2e20b9b9d6f6a125827a98495ef4