Bug 1715944 - Set the cached theme handle to Nothing() if the holding value was null.  r=Gijs
In the class stlye of Win7, `OpenThemeData` returns null.  The problem is
`ThemeHandle::Close()` did not reset the cached value to `Nothing()` if
the holding value was null.  As a result, we never fetch the theme handle
after the theme is changed.

Differential Revision: https://phabricator.services.mozilla.com/D121288

UltraBlame original commit: aeb33cef0825b4a54003fa7acef2025dc12da55e