Bug 1710524 - Fix intermittent failure in testWeakMap_keyDelegates r=sfink
The problem is that we are waiting for an incremental GC to get past the
prepare phase, but sometimes it ends up finishing completely, which causes
finishGC() to assert. The fix is to be more careful about checking before
calling that.

Differential Revision: https://phabricator.services.mozilla.com/D121393

UltraBlame original commit: ca9b5386b47d10cfa351da0de8705acf391a360d