Bug 1341265 - Part 8: Optimise Set.prototype.has for strings in CacheIR. r=iain
Inline `Set.prototype.has` in CacheIR when called with strings.

Normalising a string input is a bit annoying, because we have to atomise the
string, which means performing a VM-call. And in contrast to other VM-calls,
we have to continue the CacheIR operation after the call, which means we can't
implement this as a shared method in `CacheIRCompiler`, but instead need to
implement it in `BaselineCacheIRCompiler`. And furthermore we require that
atomising a string won't GC, because we spill the `SetObject` on the stack.

Differential Revision: https://phabricator.services.mozilla.com/D118974

UltraBlame original commit: 3fa14377770a1da734d30b0000ffa15c19c50896