Bug 1719194 part 5 - Reorganize the previous parse cache to hide internals. r=arai
This patch does 2 things:
 - It moves the closedOverBindings out of the cachedGCThings_.
 - It hides the ScriptIndex extracted from cachedGCThings_.

The cachedGCThings_ used to contain both the TaggedParserAtomIndex and the
ScriptIndex as a TaggedScriptThingIndex. These 2 indexes are referring to data
which are not the same sources. The TaggedParserAtomIndex are indexes in the
parseAtom table, while the ScriptIndex are either newly allocated or reference
to an existing Stencil which is used as input.

Thus, the closedOverBindings_ are moved away of the cachedGCThings_ to highlight
that these are not indexes into an existing Stencil, but references to the
CompilationState.

As the ScriptIndex contained in the cachedGCThings_ are not valid for the
FullParseHandler, the ScriptIndex returned value is now hidden under private
functions, and the function index is used as a replacement. This way, the
ScriptIndex reference should not leak in the FullParseHandler, thus avoiding
confusion.

Differential Revision: https://phabricator.services.mozilla.com/D120635

UltraBlame original commit: dc6a883c109022352e4caab5053c93b5d4b48859