Bug 1723741 - Fix how SVG images in blob recordings would always be transparent. r=gfx-reviewers,jrmuizel
This occurred due to a signature mismatch with
gfxUtils::DrawPixelSnapped. A previous patch in bug 1712855 put the
wrong field in the opacity parameter. It implicitly converted a bool to
a float silently, causing the opacity to be 0.0.

Differential Revision: https://phabricator.services.mozilla.com/D121628

UltraBlame original commit: fa3677b2a1e568a0831c7f0a1507cd9972bdd420