Bug 1721286 [wpt PR 29712] - [LayoutNG] OOF with inline CB relative offset, a=testonly
Automatic update from web-platform-tests
[LayoutNG] OOF with inline CB relative offset

This CL started off as a follow-up to CL:2995308. CL:2995308 had
added support for handling relative offsets for inline containing
blocks when inside a multicol, such that the relative offset is
applied to the OOF descendants after fragmentation has taken place.

There was a concern around writing mode conversions, so in attempt
to find any problems, I updated the tests added in CL:3025848 to
apply a relative offset to inline CBs in various writing mode
scenarios. No problems were discovered with the property added in
CL:2995308, but a more general bug was discovered, unrelated to
fragmentation.

The bug that this CL addresses is specific to when a relative
offset is applied to an inline CB when the writing mode is
vertical-lr. The problem appeared to be with the fact that
ComputeRelativeOffsetForInline() reverses the offset direction
in vertical-lr. This offset is then added to the static position
of the OOF descendant, resulting in the OOF shifting in the
opposite direction than expected.

To fix this, remove any relative offsets from the OOF inline
ancestors when calling PropagateChildData() to avoid the incorrect
relative offset being added to the static position. Instead, a new
method was added, called ComputeRelativeOffsetForOOFInInline(), that
will calculate the relative offset to use when applying to an OOF
descendant.

An accumulated relative offset for OOFs is now calculated in
NGInlineLayoutStateStack::ApplyRelativePositioning() using
ComputeRelativeOffsetForOOFInInline(). This accumulated offset is then
applied to the OOF descendant's static position when it is
first discovered. It is also applied to the inline relative_offset
stored in OutOfFlowPositionedNode when the OOF's inline_container is
discovered (so it can be applied after fragmentation has taken
place).

Note: ComputeRelativeOffsetForOOFInInline() still reverses the writing
mode in the case of RTL. I'm not sure why the reversal is expected
in one case and not the other, so I'm open to ideas on why this
might be or if there may be a better approach to addressing this
bug.

The various writing mode tests for OOFs, both with and without a
multicol, were updated to add a relative offset to the inline CBs
since this bug appeared in both cases.

Bug: 1231279
Change-Id: I40038524a70e6eb191790c988b24b476fee2d56e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3039319
Commit-Queue: Alison Maher <almahermicrosoft.com>
Reviewed-by: Koji Ishii <kojiichromium.org>
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Cr-Commit-Position: refs/heads/master{#907620}

--

wpt-commits: dd695beae0f59651ccbbb26b9c6f73476d643cfe
wpt-pr: 29712

UltraBlame original commit: 23ec5e0e0b704a1195e40a35f37d1f7bd74fdc2f