Bug 1722389 [wpt PR 29793] - [LayoutNG] Nested fixedpos in multicol w/ viewport CB, a=testonly
Automatic update from web-platform-tests
[LayoutNG] Nested fixedpos in multicol w/ viewport CB

There were a couple of cases that resulted in a prepaint crash for a
nested fixedpos in a multicol whose containing block existed outside
the multicol.

1. When walking a missed OOF in prepaint, we skip over any OOF children
so that they can be found during fragment traversal. However, if that
OOF child is a fixedpos, and its CB is outside the multicol, we will
miss walking that fixedpos in fragment traversal.

To fix this, keep track of any fixedpos elements that we could miss in
this way in |pending_fixedpos_missables_|, and walk those after we
finish walking the multicol.

2. If a fixedpos is nested in a non-missed OOF, and its CB is outside
the multicol, we could miss walking it in prepaint when we check
if the fragmentainer is the first for node (ie. we'll skip the fixedpos
if the OOF parent starts in fragmentainer that is not the first).
Instead, check if the parent is the first for node to make sure that we
walk legacy-contained OOFs exactly once.

Bug: 1200328
Change-Id: I623afe38eef6c2abf827ed75ee78935adc7052ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3050673
Reviewed-by: Xianzhu Wang <wangxianzhuchromium.org>
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Commit-Queue: Alison Maher <almahermicrosoft.com>
Cr-Commit-Position: refs/heads/master{#907781}

--

wpt-commits: b3527f4e824c717e441d8e6c6e0a5aa5373e32b1
wpt-pr: 29793

UltraBlame original commit: e1d71d19690354df3b407b9f487a83e512aefea2