Bug 1711126 - The module index in the third-party modules ping may be -1. r=mhowell
This patch updates the regex pattern to match data where the module index of
a callstack is -1 or a non-negative integer.  -1 is legitimately set in
`CreateJSStackObject` when the frame address is not within any module in the
module list.

If the pattern did not match in debug build, calling `toBoolean()` hit the
assert because `matchResult` was not a boolean but null.  This patch fixes
that problem, too.

Differential Revision: https://phabricator.services.mozilla.com/D122020

UltraBlame original commit: 3c5a838521b1cf4616e3ea9f1b5b3ea8769e73dd