Bug 1642516 - In the Lul stack walker, don't subtract 1 byte from the return address, for consistency with other stackwalkers. r=gerald
This makes all our stackwalkers in Firefox consistent with respect to return addresses:
For non-leaf frames in stacks, the code address now always points to the instruction
*after* the call instruction, i.e. to the instruction that will be executed once the
function returns.
For symbolication purposes, 1 byte will need to be subtracted in order to obtain correct
line number + inline stack information for  the call instruction. This subtraction will
be the responsibility of the Firefox profiler front-end, not of the stackwalkers.

Depends on D121930

Differential Revision: https://phabricator.services.mozilla.com/D121931

UltraBlame original commit: 306a7f49a334cff996f874ff56c318a71facc22b