Bug 1724727 - Make SVG image blob recordings use visibility rects. r=jrmuizel
Supplying a visibility rect to the display list builder when recording
an SVG image blob will allow the display list builder to trim out
unnecessary items and make the recordings smaller. This is particularly
important when we have an SVG image used as an atlas.

This patch also simplifies the call path for the recordings. This allows
us to avoid unnecessary clips and transforms and shrink the recording
further, as well as make the code more auditable.

Differential Revision: https://phabricator.services.mozilla.com/D122125

UltraBlame original commit: fdeba1ebcb4d3ad1eb0fa22d07b85f87b750ee3c