Bug 1724830 - Use _GLIBCXX_RELEASE in stdc++compat.cpp. r=firefox-build-system-reviewers,andi
Since the minimum required version of gcc and libstdc++ is 7.1 as of bug
1536848, the _GLIBCXX_RELEASE macro is available on all supported
versions. Which means instead of grabbing the largest version of the
libstdc++ symbols and tweaking for that, we can just tweak for
_GLIBCXX_RELEASE. We also remove the conditions that are always true due
to the versions involved.

Differential Revision: https://phabricator.services.mozilla.com/D122194

UltraBlame original commit: 3d61ee85ebb7a1c95122a13f4f9f182175aaaa24