Bug 1678895 - Add GeckoSessionTestRule.promiseAllPaintsDone to make sure all pending paints have been finished. r=geckoview-reviewers,agi
This function will be used in the next commit to make sure there's no remaining
paint requests before proceeding each test cases. This code was mostly copied
and pasted from the commit when we introduced flushApzRepaints in bug 1660357.

This `promiseAllPaintsDone` is a simplified version of promiseAllPaintsDone in
paint_listener.js.  Though the paint_listener version is supposed to be loaded
and supposed to add an event listener for MozAfterPaint events in the first
place, whereas this version is supposed to keep adding an event listener for a
MozAfterPaint event until there's no pending paint request when it gets called,
it's a compromise but it should work as expected in most cases.

Differential Revision: https://phabricator.services.mozilla.com/D122310

UltraBlame original commit: d7f0b173486f58e84fb44d82e26e84ce2d492fdf