Bug 1720650 - Fix codegen problem in narrowing.  r=yury
The ARM64 architecture uses two instructions to implement SIMD narrowing,
SQXTN and SQXTN2 (and their unsigned-saturation variants).  These are not
quite symmetric: the former will clear the upper bits of the output register
and set the lower bits, while the latter will just set the upper bits.
To implement an across-the-vector narrowing, the former must be executed
before the latter.

For some operations, we execute them in the other order to save a move.
But this is incorrect, so this patch reintroduces the move when it
is required.  This effectively makes the three-address operations the
same as the two-address operations and it's possible we want to
propagate the fix higher up in the stack to avoid the duplication of
code, but for now a local fix seems best.

Differential Revision: https://phabricator.services.mozilla.com/D122587

UltraBlame original commit: 2e3e1b631c62e990980fed64df50875de8c28600