Bug 1725002 [wpt PR 29972] - [LayoutNG] Break token for new column before spanner, a=testonly
Automatic update from web-platform-tests
[LayoutNG] Break token for new column before spanner

In CL:3018723, logic was added to the column layout algorithm to lay
out any OOF descendants when we find a spanner to ensure that any OOF
descendants found before a spanner are actually laid out before the
spanner.

If a new column was created for an OOF before the spanner, we wouldn't
end up creating a column break token for that new column if the OOF
didn't break. This CL adds logic to NGSimplifiedOOFLayoutAlgorithm to
make sure that a break token is created for the column in this case.

Additionally, if any columns are added after the spanner, we would end
up crashing because the new column created for the OOF was not
taken into account when adding the subsequent column (i.e. the
sequence number and consumed block size were incorrect, and
the subsequent column was added to the flow thread at the wrong
index).

To fix this, update the next column break token in MulticolPartWalker
to the break token of new column added for the OOF. In order for layout
to continue correctly from this new break token, any child break tokens
from the previous column needed to be copied over to the new column
generated during OOF layout. Otherwise, those children would not resume
layout after the spanner.

In addition, if an existing column fragment was replaced via
NGSimplifiedOOFLayoutAlgorithm, and it had a child that had broken due
to a break before, we would miss adding that child break token via
child_iterator_. To fix this, add any child break tokens from the old
fragment that were the result of a break before prior to simplified
layout completion.

Bug: 1237247
Change-Id: I55c73fc36a604e858dde5859f65a1ca66023e54a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3082315
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Commit-Queue: Alison Maher <almahermicrosoft.com>
Cr-Commit-Position: refs/heads/master{#910828}

--

wpt-commits: 1bce3fabaaf18dda576d83f64f3ef03bdd647e3b
wpt-pr: 29972

UltraBlame original commit: 6e2e6d13dedaaa97fb8b8dd7b8ea3ac1ee02af2d