Bug 1723200 - Revise ancestor transform assertion in WebRenderLayerScrollData::Initialize(). r=tnikkel
What we want to assert is that, apart from a couple of edge cases,
aAncestorTransformId is represented in mScrollIds.

The previous approach assumed that mScrollIds was populated entirely
in the loop above, but in fact it can also be populated via
AppendScrollMetadata().

Differential Revision: https://phabricator.services.mozilla.com/D123036

UltraBlame original commit: 24a4ff02ef897dfd4bdc638a1118556a48955097