Bug 1697901 - [network markers] Support channel cancelation r=valentin,gerald,necko-reviewers
This patch supports channel cancelations with network markers.

Because of all the possible ways to stop a channel and hence all
possible end markers, it also introduces a new boolean to reduce the
complexity. Indeed it happens that 2 "end functions" are called, but we
need to insert only 1 end marker to have a well-formed result in the
Profiler frontend.

Differential Revision: https://phabricator.services.mozilla.com/D122111

UltraBlame original commit: fddd1f768e106f66525f002c211df992da982caf