Bug 1725704 - Add test to cover CDM protection query. r=alwu
This adds coverage via the clearkey CDM to ensure that the output protection
query is working as intended. These tests ensure that the output protection
machinery recognizes when browsers windows are being captured, and this is
reflected by the clear key CDM changing key status on a test key id.

Two cases are covered:
1. Media playback is started, then during media playback a screen capture is
started and stopped. The test checks that we set restricted key status while
capture is active, and stop doing so once the capture stops.
2. Start a screen capture then start media playback, then stop capture while
playback is still active. This is similar to the first case, but the media
should start in a restricted state, as capture is ongoing when playback starts.

Differential Revision: https://phabricator.services.mozilla.com/D122784

UltraBlame original commit: 3a9c1994be60f10f06219fc19cbd707d16ccf8cf