Bug 1727328 part 7 - Use cached shape for PlainObject with default prototype. r=jonco
Because plain objects are created with different AllocKinds, we use an array of
shapes for this. A bit annoying but because plain objects are so common it's
probably worth handling all of them.

Depends on D123461

Differential Revision: https://phabricator.services.mozilla.com/D123462

UltraBlame original commit: 02f18eaee023755f4ac47d253c7e5a356e3dc0a8