Bug 1724997 - Manually choose a tag for tracing markers. r=pehrsons
It would be nice to be able to auto-generate a string, but on Windows we only
have a very long name with all the qualifiers, and it's unreadable in practice
in the marker view.

This allows having names that are more human friendly and shorter. Generally
it's going to be the name of the function though.

Differential Revision: https://phabricator.services.mozilla.com/D122737

UltraBlame original commit: ae80e44abdb29fbb06623dfaafe8e0a1f6b85733