Bug 1724031 - Part 5: Store JSFunction's JIT data / script field as a JS::Value r=jandem
In preparation for storing this in a slot, store it as a JS::Value.

All values including JSScript pointers are stored as private values becuase the
JIT assumes it can access SelfHostedLazyScript pointers and JSScript pointers
in the same way. This means we will still need a trace hook when we store this
data in slots.

Differential Revision: https://phabricator.services.mozilla.com/D123085

UltraBlame original commit: 1415c1d5b485707aa80a5141c1dfa9fc5d202430