Bug 1727177 - [devtools] Fix devtools.netmonitor.saveRequestAndResponseBodies preference with new fission codepath. r=bomsy,devtools-backward-compat-reviewers
We weren't applying the preference to the parent process NetworkEventWatcher
because we were only trying to update this via the WebConsole actor.

In addition, this was introducing intermittent failures in browser_toolbox_frames_list.js
because of the pending request made to the console actor.
Doing only one unique request to NetworkEventParent helps reducing the number of requests and intermittents.

Differential Revision: https://phabricator.services.mozilla.com/D123379

UltraBlame original commit: 142afd30347981da51659456b4075f943380bc99