Bug 1710668 - Allow Gecko to dismiss a prompt. r=owlish
This change adds a new API that allows Gecko to dismiss a prompt automatically.

This will be used to dismiss prompts when they shouldn't be displayed anymore,
e.g. when using basic auth, we hide the prompt if the server rejects the login
information, as we know that the login info is not correct, so there's no point
in saving it.

We will also dismiss prompts automatically when they become stale, e.g. if a
permission prompt is raised and then the user navigates away.

To be able to dismiss a prompt we need to keep track of existing prompt
instances for a Session. We assign each prompt a unique, randomly generated,
UUID that is used to match the javascript prompt instance with the java
counterpart.

The prompts are stored in PromptHolder. Because the PromptController is owned
by the GeckoSession, whenever the session is GC'd we will also release the
prompts associated to it (if any).

Differential Revision: https://phabricator.services.mozilla.com/D122504

UltraBlame original commit: b59e1f866e175d4a70773f76a9095d4aa4211b7c