Bug 1728451 [wpt PR 30272] - Lazyload: Lazyload images can load from the list of available images, a=testonly
Automatic update from web-platform-tests
Lazyload: Lazyload images can load from the list of available images

As per step 6 in #updating-the-image-data [1], all images including
lazyloaded ones are able to load immediately from the list of available
images. This CL adds a test ensuring this is the case, specifically for
out-of-viewport `loading=lazy` images.

If it is determined in https://github.com/whatwg/html/issues/7005 that
even in-flight requests are able to be placed in the list of available
images, this would increase the utility of the list of available images,
and maximize the number of requests that can "coalesce" with entries
in said list. In that case, we'll need to:
 1.) Correct the tentative expectations introduced in
     crrev.com/c/3133524
 2.) Write a test specifically for lazyload to verify that lazyloaded
   images don't have special behavior, and can indeed coalesce with
   in-flight requests in the list of available images

[1]: https://html.spec.whatwg.org/multipage/images.html#updating-the-image-data:list-of-available-images

R=domenicchromium.org

Bug: 994329
Change-Id: I511c09073226f729782c47cc135d36efd2ef4723
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3133207
Commit-Queue: Dominic Farolino <domchromium.org>
Reviewed-by: Domenic Denicola <domenicchromium.org>
Cr-Commit-Position: refs/heads/main{#917059}

--

wpt-commits: 57c7282839049b334bd96aa6cc7e336b2dfae558
wpt-pr: 30272

UltraBlame original commit: 23b087129de3dacd606057b6fa95e1422c61aeb5