Bug 1709421 - Abort generating partial digest of existing file when download is cancelled, r=necko-reviewers,dragana
There were a number of crashes caused due to a hang on a background
TaskQueue, and in many of these bugs there was a read from
BackgroundFileSaver on the stack. It appears that we are stuck in this
loop late in shutdown, when we are attempting to shut down TaskQueues
and XPCOM threads.

I'm unsure if we would cancel these requests already, however I noticed
that we wouldn't abort this read even if we did. This change now
performs a check to see if the BackgroundFileSaver has errored between
each read, which should cause aborts to also abort this loop, and
hopefully avoid potential shutdown hangs.

Differential Revision: https://phabricator.services.mozilla.com/D124123

UltraBlame original commit: 6d0f2956dd133643862afcbc6ae34f1ff6474369