Bug 1715275 - Add support for a JS WebAssembly.Function object r=rhunt
This introduces a new Constructor WebAssembly.Function which is part of
the WASM JS-API type reflection work. This allows normal JS functions to
be wrapped with a type and used with WebAssembly objects (e.g. Tables).

All exported WebAssembly functions will now be of this type, which nows
provides a .type() method which will expose the arguments and results to
JS.

Differential Revision: https://phabricator.services.mozilla.com/D118712

UltraBlame original commit: 3b62bd57747f21c800a93da325d999fa5efb3850