Bug 1721899 - Update waitForShutdownDecoder helper function to use mozRequestDebugInfo. r=bryce
The waitForShutdownDecoder test helper function was introduced in bug 1521964,
but the interface for the function that it was calling (mozDebugReaderData)
changed in bug 1542674. Due to the way the promiseWaitForCondition used to
work before bug 1721898, the sudden breakage of that interface was silently
ignored and this helper function stopped working without us knowing.

After bug 1721898 landed, this breakage suddenly became visible. This patch
updates waitForShutdownDecoder to use the machine-readable debug data
structure that was introduced in bug 1542674. It's still a hacky way of
determining when a video decoder has shut down, but I'm not aware of
any other way of doing that, and this is just for a test.

Differential Revision: https://phabricator.services.mozilla.com/D121766

UltraBlame original commit: 3c0107dc649de656ca423e7a5a705ba824f6ea3a