Bug 1728693. Use IsRootContentDocumentCrossProcess in nsSubDocumentFrame::BuildDisplayList. r=mattwoodrow
Not sure we even need this anymore but just doing this so I can finish the fission audit.

Added way back in bug 588407.

This shouldn't actually be any behaviour change: anything that is IsRootContentDocumentInProcess but not IsRootContentDocumentCrossProcess should be a oop if with fission (unless I'm missing a case), so we will never hit this code path, we'll hit the remote iframe path above.

Differential Revision: https://phabricator.services.mozilla.com/D124277

UltraBlame original commit: 868ad849aa0554d9d48d61bf247954b1fe34f90e