Bug 1728429 - Update the Firefox Suggest Sponsored action text by wrapping the action text. r=dao
Summary of changes:

* Remove the Sponsored action text for Firefox Suggest sponsored results
* Instead show Sponsored below their titles
* Remove the Firefox Suggest action text for non-sponsored results

This uses the action text for "Sponsored" as we do now, but it wraps it below
the title.

The Figma spec (link in Jira ticket) shows URLs in Firefox Suggest results, the
same as history results. I asked Natalie about it, and that's not correct, or at
least we shouldn't implement that now. If we do need to show URLs, then this
approach is probably worse than D124519 because wrapping the action would also
force the URL on a new line since the URL comes after it.

Differential Revision: https://phabricator.services.mozilla.com/D124563

UltraBlame original commit: d8135e9b2b8f11b741aabffad67db0b5c4febf00