Bug 1728993 - Handle same-sample case when there are no stacks - r=florian
A scenario for a thread may be that the first entry has an empty sample (usually if stack-sampling is turned off) and no running times (first-ever sample, or "cpu" is off), so it's not output, and there is no corresponding stack table entry.
Then there may be a "same sample" entry.
- If there are no running times, nothing should be output.
- If there are running times, an empty stack should be output, but for that we need to ensure that the root frame is present in the stack table entry, because that represents an empty stack.

Differential Revision: https://phabricator.services.mozilla.com/D124612

UltraBlame original commit: c045e3ee965092f3aeec845281f83ea2e26dbc98