Bug 1724650 - part 7: Make `HTMLEditor` treat preformatted linefeed characters correctly when retrieving a line range r=m_kato
`HTMLEditor` tries to collect all children in a line.  However, it does not
handle preformatted linefeed characters.  Especially when there is a
preformatted linefeed immediately before a block boundary, it becomes invisible.
So, the range of line needs to extend the range correctly similar to the case
if there is an invisible `<br>` element.

Differential Revision: https://phabricator.services.mozilla.com/D124561

UltraBlame original commit: 338e26ff2249b5960c48b42e93dbba067cff149f