Bug 1728549 - Make pointer values RegPtr. r=jseward
Traditionally pointer values (tls, heap pointer) were represented as
RegI32 because we "know" that that is really a GPR, even if the GPR is
64-bit on 64-bit systems.

Later, we introduced RegPtr to hold pointer values that were not
reference values, but used only in some cases.  Let's use RegPtr to
hold the tls and heapreg pointer values now, introducing better type
safety that will aid us in the memory64 work.

There's a wrinkle because RegPtr can't be stored on the value stack.
Recently a "RegIntptr" abstraction was introduced as part of the
cleanup of WasmBCMemory, but this is not necessary at this time: all
we need is a local conversion to RegI32 or RegI64 as appropriate for
the platform, in the one place where this is needed, to push a RegPtr
value.  So simplify this code by removing RegIntptr again and adding
some local conversion functions.

Drive-by fix: Remove some redundant annotations on method definitons.

Differential Revision: https://phabricator.services.mozilla.com/D125060

UltraBlame original commit: 41a6e728e0bed84ca8c34f440b4b38caf1c52ca6