Bug 1729458 - Part 1: Simplify DomPromiseListener, r=smaug,rpl
This makes a couple of changes, including removing the call to
AppendNativeHandler from the constructor, and using the nullable nature of
std::function instead of wrapping them in Maybe.

The main advantage of this change, however, is that it no longer acquires a
reference to `this` during the constructor, which could be unsafe.

Differential Revision: https://phabricator.services.mozilla.com/D124827

UltraBlame original commit: 704dd03288b1953b4098987a7b34796c91cbc818