Bug 1717104: Align virtualenv_name with requirements definition r=ahal
Maps virtualenvs one-to-one with their associated requirements
definition.

For example:
```
Name: "docs"
Virtualenv location: <snip>/_virtualenvs/docs
Requirements location: $topsrcdir/build/docs_virtualenv_packages.txt
```

An issue to be resolved in the future is that it's tricky to know that,
when you define a new virtualenv, you have to *know* that a requirements
definition needs to exist in `build/`.

As part of this change, the default virtualenv ("common") was
split from the build virtualenv ("build").
This required changes to the python-test virtualenv since
python-tests depend on `glean_parser`, but were getting it
implicitly from the "build" virtualenv's requirements.
This addition to the `python-test` virtualenv is temporary and
will be removed when bug 1724273 is resolved.

Differential Revision: https://phabricator.services.mozilla.com/D122891

UltraBlame original commit: 9ec8241c2e1a008adffe0c5e1fd30b2e47601e78