Bug 1730759: Remove unreachable return statement from KeyEventHandler::KeyToMask. r=masayuki
This return statement is just a copy of the 'default' case in the preceding
switch statement.  It's redundant and unreachable because every case in the
switch statement has a return statement.

clang correctly warns that it's unreachable, so this patch fixes a clang build
warning.

Differential Revision: https://phabricator.services.mozilla.com/D125595

UltraBlame original commit: a485fe15f675ad3ca5ade2e150760d7125fdd417