Bug 1639153 - Introduce indirect stubs to optimize call_indirect. r=lth
This patch introduces indirect stubs.
An indirect stub is a stub that takes care of any switching activities needed for call_indirect.
Before this patch, we have to conservatively assume that any call_indirect's target can be from a foreign instance,
but now we don't need to do this because now Table[i].code is replaced with the address of the corresponding indirect stub.

There is one optimization - if we 100% sure that call_indirect's target belongs to our instance then we can omit indirect stub
and use raw function call with signature check of course.
We 100% sure that call_indirect's target is the same instance iff Table is private and we don't export it.

However, even if a table is exported but not used we can still use the optimization but this logic for some next patch.

Differential Revision: https://phabricator.services.mozilla.com/D117123

UltraBlame original commit: d6506b5ee07b5c65f889ce7c2214b0172f205898