Bug 1730707 - Skip the Staging CPU buffer during uploads in some cases. r=gfx-reviewers,lsalzman
On windows we first group images into 512x512 staging CPU buffers which are uploaded via glTexSubImage2D (UpdateSubResource). When something is almost as large as the staging CPU buffer, we are unlikely to be able to batch it with other items. In this situation we are better off skipping the CPU buffer and uploading off of the image data directly.

Differential Revision: https://phabricator.services.mozilla.com/D125564

UltraBlame original commit: 16892a0df768d8f31e70323eb58e6ba5c16745fb