Bug 1730876 - Never discard quick suggest results and dedupe other URL results with a lower or same prefix rank. r=mak
This makes a couple of changes:

* Never discard quick suggest results. At some point we may want to revisit, but
  right now we should always show them.
* Discard other URL results whose URLs dupe the URLs of quick suggest results or
  whose URLs have a lower prefix rank.

The first change by itself fixes the bug, but without the second change, we'll
still show the eBay history result above the eBay quick suggest, which I
confirmed with Natalie is not what we want. We only want to show the quick
suggest.

Keep in mind that the results in this case have the same URLs, same prefixes.
When a non-quick-suggest result has a *higher* prefix rank, I think we should
*not* discard it. In that case we should show both the quick suggest with its
lower prefix rank and the other result with its higher rank.

Differential Revision: https://phabricator.services.mozilla.com/D125768

UltraBlame original commit: 54b20adae1935f8057318be0d5b1abf038f90e97