Bug 1700580 Part 2 - Add a helper class to cache flex item's metrics in its most recent reflow. r=dholbert
This patch introduces a class that caches metrics in flex item's final
reflow. With it, we have a better symmetry for caches, one for measuring
reflow and the other for final reflow. Also, we don't have to explain
that mLastReflowTreatedBSizeAsIndefinite is only meaningful if
mFinalReflowSize is set.

Differential Revision: https://phabricator.services.mozilla.com/D124967

UltraBlame original commit: 45519af7e8d2b746f0b0fc495228f438b10c1572