Bug 1727254 - Simplify i32 multiply in baseline compiler. r=yury
The baseline compiler previously assumed that the multiplication
instruction that was emitted for mul32() was the very old-fashioned
32x32->64 instruction that requires the input in eax and clobbers
edx.  This is not so; the chip has a general GPR1 *= GPR2
instruction that has none of these problems when using the two-argument
form of mul32().  So simplify.

Differential Revision: https://phabricator.services.mozilla.com/D125961

UltraBlame original commit: a328c1e9d66b297a03c73d4be9a5ec75ecb410f7