Bug 1653417 Part 3: Convert NativeLayerCA::Representation flags to use a bitfield. r=gfx-reviewers,mstange
The size of the NativeLayerCA::Representation struct seems to be a factor in
the power consumption of native layer updates. Converting this series of bools
into a bitfield is an easy way to reduce the structure size.

Differential Revision: https://phabricator.services.mozilla.com/D124992

UltraBlame original commit: 6bb1dd91e24be029e0b9b09a07efe023d352d22d