Bug 1725524 - Expect no stacks in the loading events of blocked modules.  r=mhowell
When a module is loaded, `NotifyEndDllLoad` captures its stack back trace only
when the loading result is success.  In GTest, if a batch of loading events
passed to `ValidateUntrustedModules` consist of blocked modules, the count of
modules in the stacks is zero.

Differential Revision: https://phabricator.services.mozilla.com/D126040

UltraBlame original commit: 094c28cf5863dea0b1f456f3ba6c7fc2784bb2bc