Bug 1731218 - Part 0: Remove unnecessary template from Rooted constructor r=sfink
I don't think this does anything. The type is always present so we won't
restrict use of this constructor by SFINAE, although I may have misunderstood
the purpose of this.

I tried 'fixing' the SFINAE to what I though it should be by using enable_if_t
but it had the same effect as removing it. In both cases the code compiles, and
if I try to construct a rooted with an inappropriate argument it fails with a
sensible error.

Differential Revision: https://phabricator.services.mozilla.com/D125949

UltraBlame original commit: 5c9145f71f553a5d739ce32abde81d2e7479ee15