Bug 1731793 - [devtools] Fix node picker error when closing browser toolbox. r=ochameau.
onInspectorFrontDestroyed was throwing because when it
is called from the destroy callback of watchFronts, it
doesn't get a second argument, and the destructuring fails.
This patch defaults the second parameter to an empty object
in order to avoid this.

Differential Revision: https://phabricator.services.mozilla.com/D126235

UltraBlame original commit: 0cab7f09323ccee14f3eaa86d42ba154137feb8c