Bug 1729154 [wpt PR 30348] - [cc] Recreate high resolution tiling w/o text when scale is not ideal, a=testonly
Automatic update from web-platform-tests
[cc] Recreate high resolution tiling w/o text when scale is not ideal

CanRecreateHighResTilingForLCDTextAndRasterTranslation used to only be
relevant when the raster transform changed its translation. The scale
was 1D and equal to the tiling key, so it would always be ideal. Then it
was fine to not recreate high resolution tilings if there was no text.

However, in r872117 I added support for 2D scales, defining the key as
the maximum component. Then it can happen that we find a previous tiling
with the given key, but with a different scale.

In that case, reusing the tiling just because it has no text can produce
very poor results. So this patch changes that case to only reuse if the
scale is ideal.

Also, CanRecreateHighResTilingForLCDTextAndRasterTranslation is renamed
to CanRecreateHighResTilingForLCDTextAndRasterTransform since it takes
the scale into account.

Bug: 1242860

TEST=external/wpt/css/css-transforms/transform-3d-scales-different-x-y-dynamic-002.html

Change-Id: Ia68f90728a170f864fb6c0ff71b831293b887c71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3141874
Reviewed-by: Xianzhu Wang <wangxianzhuchromium.org>
Commit-Queue: Oriol Brufau <obrufauigalia.com>
Cr-Commit-Position: refs/heads/main{#918387}

--

wpt-commits: a747e0292b02826258bcbfb3ab9528c60d527018
wpt-pr: 30348

UltraBlame original commit: 709d34d03f45babebaa111d03beab66f660f9781