Bug 1337723 - wasm: Report tier-2 warnings and errors to stderr. r=lth
Tier-2 tasks are finished on a helper thread and don't have a
JSContext* to report errors/warnings to. It may be useful for
end-users or engine developers to be notified when a tier-2
task fails and why. This commit implements logging of errors
and warnings by printing to stderr.

An alternative approach would be to get a runnable on the
originating JS thread that reports the errors/warnings using
that *JSContext. I'm not sure the best way to do this, so I
went with the simpler approach for now.

Differential Revision: https://phabricator.services.mozilla.com/D125585

UltraBlame original commit: ce92432790779f544502ef93e87f6b4b38579e68