Bug 1731223 - Move wine detection to toolkit/moz.configure. r=firefox-build-system-reviewers,andi
As of now, wine is used with:
- fxc, which is detected in toolkit/moz.configure
- midl, which is detected in toolkit/moz.configure
- ml64 or armasm64, for some assembly sources that aren't used in
  projects that don't include toolkit/moz.configure.

We can thus move it there.

As of bug 1731195, there is no remaining use of wine for artifact
builds, so we also make it limited to builds with a compile environment.

Differential Revision: https://phabricator.services.mozilla.com/D126166

UltraBlame original commit: 59a15d9c5391034a1d92192ba1721044a202e91f