Bug 1718444 - Convert nsIUpdateCheckListener to async r=nalexander,application-update-reviewers
I want to be able to check if the update-available notification is fired during testing. The entry point to firing that notification is UpdateService.onCheckComplete, but it currently kicks off its work and does not wait for it to finish. This means that I can wait for the update-available notification to be fired, but I cannot easily wait for it NOT to be fired, which is what I want to be able to test here.

It looks though like it should be easy enough to just convert this interface to an asynchronous one. This will make it much easier to call onCheckComplete and know by the end of it if the update-available notification gets fired or not.

Differential Revision: https://phabricator.services.mozilla.com/D126162

UltraBlame original commit: d48f09850f941f9fdac73cb5e62496fbb120a092