Bug 1729861 - Fix test for bug 418986. r=jfkthame
The test was relying on media queries not being recomputed, because it
ran something like this:

  1. RFP off: add a bunch of images with class "testImage" to the
     document. Then test all images with class "testImage".

  2. RFP on: do the same.

Note that there's no "remove all images added in step 1" step between 1
and 2. So the test with rfp on was also testing the images from step 1.
This worked before because resistFingerprinting didn't trigger
recomputation of the previous images.

Differential Revision: https://phabricator.services.mozilla.com/D126152

UltraBlame original commit: 129487f35030f906f562878e07a60dd47edd8eec