Bug 1729653 - Make `mozInlineSpellChecker` handle events in the system group at capturing phase r=m_kato
Currently, `mozInlineSpellChecker` handles events in the default group.  This
means that the listener may not run if web app stops the propagation.  For
keeping the event listener order as far as possible, this patch changes to
listen to the event at capturing phase in the system group (editor handles
the events at bubbling phase in the system group).

Additionally, it listens to `keypress` events for handling caret/selection
changes.  However, they should be handled at `keydown` because those keys
are not printable.  Therefore, this patch changes to listen to `keydown`
events, but this could change the race between running the scheduled spellcheck
and following `keypress` event which is dispatched only for chrome code and
C++ event listeners.  However, this may not change it actually because the
race is changed only when the following `keypress` event delays too much.
https://searchfox.org/mozilla-central/rev/2eebd6e256fa0355e08421265e57ee1307836d92/extensions/spellcheck/src/mozInlineSpellChecker.cpp#503-504

Differential Revision: https://phabricator.services.mozilla.com/D126635

UltraBlame original commit: ca8fcd42078a2b9e2f6483b5a0088906aa4ee711