Bug 1731996 - [devtools] Fix nodePicker on remote iframe padding. r=ochameau.
With Fission enabled, hovering the padding area of an iframe would not cause the
iframe element to be highlighted.
That's because we were deliberately ignoring any event on remote frames, as we
thought the sub-document picker would take care of this.
We fix this by only checking if the event was fired on a window managed by the
target actor.

An existing test is modified to make it a Fission test case, and a few helper
for highlighterTestFront had to be modified.

Differential Revision: https://phabricator.services.mozilla.com/D126472

UltraBlame original commit: 84303236bccb30ea48422ab49f12ca0db3950ea2