Bug 1733046 - [devtools] Hide Geometry highlighter on various user interaction. r=jdescottes.
The code was written to hide the geometry highlighter when:
- the nodepicker was selected
- another node was selected in the markup view
- a node was hovered in the markup view

but this wasn't working as intended, as the line setting
the first of the event listeners was throwing, and we
don't have tests for this.

This patch fixes the faulty line and adds a test to check
that the geometry highlighter does hide on those actions.
We take this opportunity to use an AbortController to handle
the removal of the event listeners added when starting the
geometry highlighter.

Differential Revision: https://phabricator.services.mozilla.com/D126885

UltraBlame original commit: f64a4714456af90a375bd33fd4edd512e84d30a5