Bug 1733919 - Block OES_EGL_image_external_essl3 on Mesa versions prior to 20.0 r=gfx-reviewers,jrmuizel
Webrender encounters a shader compilation error when rendering video
in firefox for android on intel based chromebooks. The error is due to
a missing definition of textureSize(samplerExternalOES), which should
be defined when using the OES_EGL_image_external_essl3
extension. However, these devices are running Mesa drivers, and due to
a Mesa bug affecting versions 19 and earlier this was not defined.

We already have a fallback path for devices which do not support
OES_EGL_image_external_essl3: we use the ESSL 1.0
OES_EGL_image_external extension instead. This patch detects when we
are running on Mesa 19 or earlier, and forces us to use the fallback
path.

Differential Revision: https://phabricator.services.mozilla.com/D127418

UltraBlame original commit: 0c8983f9c20e68dea55edaed28281a8c6c47220e