Bug 1733523 - [devtools] Handle moving node selection when selected node's target gets destroyed. r=jdescottes.
Toolbox's Selection instance listens for `mutations` walkerFront event in order to detect if the selected node
was removed from the DOM tree, and in such case, emits a `detached-front` event that can be consumed by
the markup view.
But, when an iframe gets removed and EFT (or Fission for a remote iframe) is enabled we don't get the `mutations` event.
To fix this, we add a `onTargetDestroyed` method to `Selection`, that we call from the toolbox.

This fixes the browser_inspector_delete-selected-node-* tests when EFT is enabled.

Differential Revision: https://phabricator.services.mozilla.com/D127224

UltraBlame original commit: a26075e0ca2211c18798363eefd007c6650f9d89