Bug 1733025. Convert coords correctly when sending cmd_lookUpDictionary (macOS lookup) to child process. r=hiro
this.browser in toolkit/actors/ControllersParent.jsm is the top level browser, ie the browser holding the root content document. So the conversion that happens in that file converts the coordinates to be relative to the root content document, but they need to be relative to the root of whichever child process we are sending the event to.

The best way I found out how to do this was to pass the coords down to the child process still relative to the parent process widget and then in the child process use the child to parent transform matrix to make them relative to the root widget in the child process.

I needed a new nsIDOMWindowUtils functions because I don't think there is anything existing to do this.

Differential Revision: https://phabricator.services.mozilla.com/D126861

UltraBlame original commit: 9060eae95f294d4e2c2526c56f93d05e9892bbf1