Bug 1705360 - "hide" NSS DBs from meddling third party software r=jschanck,bbeurdouche,perftest-reviewers,sparky
Some crash reports appear to be indicating that initializing NSS' certificate
and key databases is taking on the order of minutes in some cases, which is
unexpected. One hypothesis is that third-party software is opening these DBs at
the same time that NSS is operating on them, causing contention and thus
slowness. This patch experimentally (in Nightly only) renames these DBs in the
hopes that third-party software might not recognize them as the DBs it's
looking for, and will thus leave them alone.

Differential Revision: https://phabricator.services.mozilla.com/D126028

UltraBlame original commit: b5a5e31457dc1d3bb674ad1300ffbd56e7abd12c