Bug 1732829 - Propagate popup position / size change after widget moves, not before. r=hiro
Bug 1696718 landed a fix to propagate the position change from layout.
However that's not correct, because the widget isn't resized until
nsView::ProcessPendingUpdatesForView is resized (and even that in some
platforms it might be async).

So the right place to propagate the position change is in
nsXULPopupManager (which we call into from the view system which listens
itself to the widget).

Let's try to enable the test for that bug everywhere with this fixed.

Differential Revision: https://phabricator.services.mozilla.com/D127801

UltraBlame original commit: 76a0d606b7557899893ec3153f74b8d369a6b555