Bug 1733938 - mTRRSkipReason telemetry should never be TRR_UNSET in telemetry r=nhnt11,kershaw,necko-reviewers
After bug 1719135 we no longer call TrrLookup from nsHostResolver::NameLookup
if we have not TRR service or if Enabled returns false. As a consequence
we'd record the TRR skip reason as TRR_UNSET - which we shouldn't do, as that
data point is useless.

This patch introduces a new TRRServiceEnabledForRecord method to
nsHostResolver that properly sets the skip reason when the TRR service
is not available or the request should skip TRR.

Differential Revision: https://phabricator.services.mozilla.com/D127442

UltraBlame original commit: a0adc6ddb2b08949afb6bec8c6be4f59f11610d0