Bug 1733335 - Make std::vector an opaque type in the profiler Rust API r=emilio
This was causing asan and dt failures because bindgen was computing the size of
MarkerSchema as 122 bytes instead of 144 bytes. This was causing a heap buffer
overflow. After making the std::vector an opaque type, it computes the size
properly as 144 bytes.

Depends on D127114

Differential Revision: https://phabricator.services.mozilla.com/D127956

UltraBlame original commit: 5640485486b012dbe5b0234549e279b5bb40caf3