Bug 1734262 - Partially fix broken test for pingsender url allowlist. r=gsvelto
The submission was always failing regardless of host because it was trying to
submit a ping that had already been deleted. Separating this segment out
into its own test makes it easier to spot to mistake.

Unfortunately, even with this fix, the test doesn't properly fail when
the logic-under-test is missing, because the connections to example.com
cause gecko to crash during testing (because MOZ_DISABLE_NONLOCAL_CONNECTIONS is set),
and the process machinery doesn't distinguish a crash from an exception.
I haven't thought of a good solution to this — suggestions welcome — but
I think it's probably not worth spending more time on.

Differential Revision: https://phabricator.services.mozilla.com/D127760

UltraBlame original commit: 1e4cafb39d62b8fb19d58fc0444a0e08d5158484