Bug 1735574 - Simplify meta viewport handling. r=smaug
We only need the whole array to deal with removals, but other browsers
don't deal with them, see:

  https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/html_meta_element.cc;l=479;drc=439c6ac35199f3828faf0f0d070d84109c9bf6a1

Given <meta name=viewport> is weird enough as it is, we should probably
do this and simplify the code.

The header data thing was just synonym to "has a meta viewport", so
simplify that code too.

Differential Revision: https://phabricator.services.mozilla.com/D128390

UltraBlame original commit: 86e71bf1ba6cef915ac48ccdcf7404c0f607d420