Bug 1721208 - Added a preprocessor #define for the CPU architecture in created .res files to resolve a build breakwith with Windows SDK 10.0.22000.x. r=mhentges,glandium
On SDK versions older than 10.0.22000.x we defaulted to 128, but with
this update, Microsoft removed the default and created a compiler error
if the CPU architecture was not specified. This change explicitly
defines the CPU architecture, resolving the compiler error.

Differential Revision: https://phabricator.services.mozilla.com/D128509

UltraBlame original commit: 53b33618fc5396106f6572855d05d5f9ee960258