Bug 1727084 - Memory64 - Huge-memory status depends on index type. r=yury
Whether a memory is huge depends not just on whether huge memory has
been enabled in the process, but whether the memory's index type
allows the use of huge-memory optimization.  It also depends on
whether the input is wasm or asm.js.  (And in the future it may depend
on eg whether the memory, if it is i64, has a declared maximum.)

Together, this fundamentally means that the is-huge-memory attribute
cannot be computed before compilation starts but must be computed once
the memory has been declared.

The difference between asm.js and wasm, and the coming of large
"non-huge" memories, also means that a memory object must carry an
attribute as to whether it is huge or not.  It cannot be determined
from the process's huge memory setting without knowing the input type,
and it cannot be determined from the size of the memory.

To make this simple, we parameterize IsHugeMemoryEnabled() by the
index type, and plumb the index type through the code where we need it
(not too many places).  Features::hugeMemory can now be removed.

We also add an IsHuge attribute to WasmMemoryObject to carry that bit.
Structured cloning must transmit that bit too.

Test cases are in the last patch in the queue.

Differential Revision: https://phabricator.services.mozilla.com/D126186

UltraBlame original commit: 5315e548b65157bccf7a1d81f82c33dd9ae1461b