Bug 1727084 - Memory64 - Allow larger-than-4GB allocations. r=yury
This is almost all plumbing: Parameterize the allocation path on the
memory's index type.

The most substantive change is to wasm::MaxMemoryPages in
WasmMemory.cpp, which now has logic to compute the right value from
the index type.  Otherwise that file was changed only by rearranging
it, which was required to get the scoping right.

Test cases are in the last patch in the queue.

Differential Revision: https://phabricator.services.mozilla.com/D126187

UltraBlame original commit: d13d8c76693acc79e9245ec6e54fcbf0f24a005f