Bug 1727084 - Memory64 - Non-atomic-RMW baseline operations. r=yury
Implement scalar, SIMD, and atomic loads and stores for the baseline
compiler for x64, x86, arm64, and arm32.  Atomic RMW operations and
wait/notify are in the next patch.

Largely, this happens through templatization and some type-specific
helpers.

Introduce 64-bit bounds checking masm abstractions.  On 32-bit
systems, heaps are limited to 2GB and bounds check limits all fit in
32 bits, so a 64-bit check amounts to checking that the high word of
the pointer is zero and then performing a 32-bit check on the low
word.

MIPS64 and Cranelift are not supported for memory64.

Test cases are in the last patch in the queue.

Differential Revision: https://phabricator.services.mozilla.com/D125183

UltraBlame original commit: 6089b4360edcfb3a4bda506e37be8ca75a9f15a6