Bug 1736353 part 1 - Remove barrier after megamorphic property gets. r=iain
We added a speculation barrier after calls into C++ from ICs (bug 1444473).
In hindsight, it seems we were overly cautious with the calls for megamorhic property
gets because the slot number comes from the shape's property map. This makes it very
different from the typical Spectre v1 attack with a user-controlled index into a
(typed) array, where the bounds check then gets delayed/speculated.

Furthermore, this happens in C++ code that's not user-controlled and it'd need a
large speculation window that includes the return to JIT code. There we have
additional mitigations when unboxing the returned Value.

This barrier has a large impact on performance: on a simple micro-benchmark it
increases our numbers from 75 ms to 170 ms on Linux x64, more than a 2x slowdown
for one of our hottest code paths. This makes it hard to justify keeping it.

Differential Revision: https://phabricator.services.mozilla.com/D128744

UltraBlame original commit: 0735f5660d2bc3ff4fac1f7204291055baedb4b4