Bug 1736530. Enable gfx/layers/apz/test/mochitest/helper_fission_animation_styling_in_transformed_oopif.html. r=hiro
This test was disabled for webrender because of an assert. The mentioned bug is resolved and doesn't seem to mention this test at all. I pushed to try and it seems to pass now. So let's enable it.

Differential Revision: https://phabricator.services.mozilla.com/D128862

UltraBlame original commit: c8be478a2054cb3a90f874d4d2a93d4903f7affe