Bug 1736103 - fix download file extension correction test some more for the no-download-modal case, r=sfoster
On try and on my machine, after D128648 the test still times out because
for the test_broken_saved_handlerinfo_and_useless_mimetypes test, first
awaiting the `task` that gets passed means that the download has already
finished by the time we create the download finished promise.

This patch moves creating that promise before the `task` is started.
It also broadens the conditions in which we await this promise, because
in the case without file extensions and without a dialog, we were otherwise
no longer awaiting anything, which meant that the next test's check for
a finished download would resolve with the previous test's download, causing
confusion/breakage.

Differential Revision: https://phabricator.services.mozilla.com/D128916

UltraBlame original commit: 1350a6c88f940fcabfe14a5a1b9eb35619ca86b0