Bug 1735643 - Unify Urlbar box HCM behaviour. r=desktop-theme-reviewers,dao
This patch unifies the HCM colors used in the identity box, the search mode indicator, and the zoom indicator. It uses --button-bgcolor as the default background for all these elements. It uses SelectedItem/SelectedItemText on hover and active. This is similar to what we do with in-content buttons.

Differential Revision: https://phabricator.services.mozilla.com/D128484

UltraBlame original commit: 46edc3da8b6cc3fe0b5ef8bb57cfd553c92120ad