Bug 1708668 - Enable Glean testing in xpcshell tests. r=chutten
This ensures Glean is initialized in an xpcshell environment on Android.
When GeckoView is embedded the surrounding app is responsible to
initialize Glean from the Kotlin side.

This also enables the first few FOG tests to run on Android.
For now it changes the Android-specific test file.
As long as not all metric types are enabled we can't reliably test other metrics.

Differential Revision: https://phabricator.services.mozilla.com/D127420

UltraBlame original commit: 7f95960e67a0e8ca55f2b60479ac396361928eb5