Bug 1737755 - [devtools] Fix browser_inspector_highlighter-iframes_01 when EFT is enabled. r=bomsy.
When EFT is enabled, we have dedicated target highlighter actors; this patch ensures
we're retrieving those (and associated highlighterTestFront) to check that highlighters
do work in iframes.
`getNodeFrontInFrames` is modified so it copies the array we pass to it, as it
performs some mutation on it, and the callsite might use it later.

Depends on D129386

Differential Revision: https://phabricator.services.mozilla.com/D129481

UltraBlame original commit: 6b014e1d206c237e2c92cf27b0ce23f0b17ce6a2