Bug 1736859 - [marionette] Explicitly turn off 'fission.autostart' if --enable-fission is not used, r=webdriver-reviewers,jgraham
The 'fission.autostart' pref is about to change defaults which means that if we
don't explicitly disable it when '--enable-fission' *is not* passed in, we'll
actually be running with it enabled.

Eventually we should remove '--enable-fission' in favour of '--disable-fission',
but that can be follow-up work once the pref has been switched.

Depends on D129309

Differential Revision: https://phabricator.services.mozilla.com/D129399

UltraBlame original commit: edd0ea14fcca51b5b9b9ab7987fc1cdad1b807a8