Bug 1734747 - Refactor --enable-lto handling. r=firefox-build-system-reviewers,andi
Using `if len(value) and value[0].lower() == ...` is confusing because
it doesn't make it super clear what defaults are. (also, the `lower()`
is not necessary).

We also stop setting value to thin when it was originally cross, because
it prevents special-casing for cross-lto, which we do, and it currently
doesn't actually work.

We also don't need to return a full namespace when lto is disabled (and
we already don't in an existing short-circuiting case), so we can remove
a level of indentation.

Differential Revision: https://phabricator.services.mozilla.com/D129609

UltraBlame original commit: c31401e9afd0107d92ceb8bec52d8f7063b17cce