Bug 1737501 - Allow contextmenu option on Windows SystemStatusBar r=mconley
We want to be able to have a different behavior for interactions with an entry
in the Windows system tray - specifically, we want to be able to open a window
on left click, and show a menu on right click. This patch allows us to control
this with a "contextmenu" attribute on the menu passed in to `AddItem`, and
receive a "systemstatusbarclick" event in the case of a left click.

Regarding tests, I don't know that we have a good way of testing this?
Suggestions are of course welcome here.

Differential Revision: https://phabricator.services.mozilla.com/D129413

UltraBlame original commit: 6aff7acd8390f8d83eb15589a0ceb376cedb2416