Bug 1654112 - Report video frame-/bitrate telemetry once per call instead of every second. r=ng
This is similar to the behavior prior to bug 1532898, which introduced telemetry
reporting every second. There's a bit of a difference in that we now only report
data if we had collected some through the 1 second interval telemetry timer in
PeerConnectionCtx.

Differential Revision: https://phabricator.services.mozilla.com/D102951

UltraBlame original commit: a55a718715f69cba1ff9fbae4787881b4d930907