Bug 1654112 - Don't register/unregister conduit with the call wrapper in ctor/dtor. r=ng,bwc
Doing this outside of the dtor, especially, removes the requirement of the
conduit having to be destroyed on the worker thread (main).

Not having threading requirements on the conduits' dtor means we can constify
conduit members everywhere, so this patch tries to do that, for added compiler
guarantees.

Differential Revision: https://phabricator.services.mozilla.com/D102288

UltraBlame original commit: ceed49af0fad6ee87f64a89adb84c791a74ced14