Bug 1654112 - Remove overzealous assert. r=padenot
This assert was intended for when an input stream is started with processing on.
However, when an input stream is started with processing off, then later when
processing is turned on it triggers this assert because mSegment contains data
from InsertInGraph().

Differential Revision: https://phabricator.services.mozilla.com/D102577

UltraBlame original commit: c8015b9110fdd2e785e2de6308c73bb8cf1505fe