Bug 1654112 - Move AudioConduit from a Mutex to an RWLock. r=padenot
AudioConduit will try to grab the lock on the audio thread, and if it fails
there is a risk of glitching. That is fine as long as the only other usage of
the lock is related to renegotiation where the audio stream is being recreated
anyway.

Then GetRtpSources was added, which, in order to not have to continuously push
(every 10ms) data from the call thread to main, grabs the lock on main thread.
With a Mutex there is now a risk of glitching audio even when we are not in
renegotiation.

A Read-Write-Lock solves this, because only renegotiation would lock the write
lock.

Differential Revision: https://phabricator.services.mozilla.com/D124368

UltraBlame original commit: 336f58a54bd8128c47fa783a508877096ba84918