Bug 1420493 - Apply maxWidth and maxHeight when processing frames, rather than on resolution changes. r=dminor
An alternative approach would be to merge max_fs with maxWidth*maxHeight and let
the VideoAdapter decide on the resolution entirely. Then the logic could have
stayed in place.

But the VideoAdapter has another algorithm for finding the ideal resolution,
which for the unittest for this feature selects 960x640 rather than 1080x720.

This moves the checking so we can keep using the same logic as before.

This also enables a disabled test so we have test coverage of these parameters.

Differential Revision: https://phabricator.services.mozilla.com/D105778

UltraBlame original commit: c674f2f606beb822a5660fae70370bb63a649c73