Bug 1738228 - Part 3: Don't use MOZ_MAKE_MEM_UNDEFINED on the start of nursery chunks r=sfink
We use MarkPagesUnusedHard/MarkPagesInUseHard on unused pages in nursery chunks
to allow the OS to make use of this memory without releasing the address space.

The latter function is currently applied to the start of the chunk including
the header, even though this part is never passed to MarkPagesUnusedHard. Since
MarkPagesInUseHard uses MOZ_MAKE_MEM_UNDEFINED on the address range it
receives, valgrind warns us about accessing memory that has been marked as
undefined when we touch the chunk header.

The fix is to not call MarkPagesInUseHard on the first page of the chunk, since
MarkPagesUnusedHard is never called on this region either.

Differential Revision: https://phabricator.services.mozilla.com/D129815

UltraBlame original commit: 1e1f79e7a3dfc461dbc7401172ee65234b772a12