Bug 1686137 - Replace 'wget' with 'requests' in install_mobile_android_sdk_or_ndk to remove the external dependency on 'wget' r=mhentges,firefox-build-system-reviewers g=firefox-build-system-reviewers
Downloading, as well as resuming a partial download, of an SDK/NDK was
implemented. The progress bar was implemented via tqdm.

This is not a generic solution, but the same ideas could be used and improved
to replace 'dlmanager' in the future.

Differential Revision: https://phabricator.services.mozilla.com/D129063

UltraBlame original commit: 3bf13dec098dbc99554ae63b4b5d98580c5c8f4e