Bug 1738636 - Return true if the browsing context is not available for a document in AntiTrackingUtils::IsThirdPartyWindow(). r=pbz
The browsing context of a document might be a nullptr in some rare
cases. This patch makes sure we won't pass a null browsingContext to
IsThirdPartyContext() in AntiTrackingUtils::IsThirdPartyWindow().

We don't have a good way to test this. We can only first land the patch
and see if the crash report stops.

Differential Revision: https://phabricator.services.mozilla.com/D130357

UltraBlame original commit: 1ae3328849542f9a1a824300c71b2f4f3ecba481