Bug 1131996 - Part 2: Support out-of-bounds read access on arguments objects in CacheIR. r=iain
The `ELEMENT_OVERRIDDEN_BIT` flag is set whenever any element is defined on an
arguments object, irrespective of whether the element is in-bounds or out-of-bounds.
That means that flag can also be used to determine if an arguments object has any
elements besides the frame arguments.

When reading a possible out-of-bounds index, we can therefore use the following
approach:
1. Fail whenever `ELEMENT_OVERRIDDEN_BIT` is set.
2. If the index is in-bounds:
  a. Return the in-bounds element unless it's `FORWARD_TO_CALL_SLOT`.
3. Else,
  a. Fail if the index is less than zero.
  b. Return `undefined`.

Plus a prototype guard check to ensure the element isn't present on any object
of the prototype chain.

Differential Revision: https://phabricator.services.mozilla.com/D129620

UltraBlame original commit: c31e81d4c5c99d698a8920e24325108b9aabf6e1