Bug 1739500 - De-duplicate "CompositorScreenshot" marker types - r=canaltinova
There were two structs claiming to be "CompositorScreenshot", which is not allowed. It was working because the JSON output was compatible, and handled in a special way in the front-end.
But it's better to combine the two structs into one, for consistency with other markers, and lower maintainability risks.

Differential Revision: https://phabricator.services.mozilla.com/D130435

UltraBlame original commit: fc15a8a590a22a6f14f6684bbcfcf85ce65a0022