Bug 1722899 - Send TouchCancel event just before dragging window r=stransky
`gdk` takes `touch-move` and `touch-end` events when dragging the window.
As a result, the touch-start event was propagated, but not touch-end. As a result, all additional touches were considered part of the same gesture as the touch that started moving the window! This affected some Linux distros with the environment variable MOZ_USE_XINPUT2 set to 1.
This patch fixes that by sending a TouchCancel event just before we transfer the touch gesture to GDK.

Differential Revision: https://phabricator.services.mozilla.com/D129349

UltraBlame original commit: 36dba57d3b902ff6a1b851afccf7b4323637821a