Bug 1731982 - Part 11: Set the `IsThirdPartyContextToTopWindow` flag to the main worker script loading channel. r=asuth
When loading the main worker script, we will need to get the clientInfo
for the loading channel. However, the loading channel is created in
content processes and the `IsThirdPartyContextToTopWindow` won't be set
in this case. So, we will get an incorrect clientInfo with a wrong
foreign patitioned principal.

To resolve this, we have to set the flag in the content processes so
that we can get a correct clientInfo. The flag comes from the
workerPrivate which reflects the fact if the worker was created in a
third-party context. And the loading channel should have the same flag
as the document which creates the worker.

Differential Revision: https://phabricator.services.mozilla.com/D129060

UltraBlame original commit: 347ce2439936076d42435aaf1b2f999e72dd0e6a