Bug 1737673 - Support 'composite' variants, r=jmaher
Often we have two variants that we want to test together. The only way to do
this currently is to create a 3rd variant that is the sum of the previous two.

This feature will allow tasks to define multiple variants using the '+' token.
E.g:

    variants: ["fission", "webgl-ipc", "fission+webgl-ipc"]

The third variant there will first apply the 'fission' configuration, then the
'webgl-ipc' configuration (in order).

Differential Revision: https://phabricator.services.mozilla.com/D130399

UltraBlame original commit: 829097878a30dffa836c72d59431df81b23dd862