Bug 1739933 - Make `TopLevelEditSubActionData::DidSplitNode()` handle both cases that either left or right node is a new node r=m_kato
Currently, it assumes that new node is the left node and split node is the
right node.  However, we need to make it possible to handle the case that
right node is new one for fixing bug 1735608.

Depends on D130458

Differential Revision: https://phabricator.services.mozilla.com/D130624

UltraBlame original commit: dcda57b3ab5803287ddad0a2f5d8526865975711