Bug 1739972 - Switch realm if necessary when creathing |this| in the NewPlainObject fallback functions r=jandem
When creating the |this| object for a scripted constructor, the shape's realm
may not match that of the context. The patch swiches realms when necessary.

This is not a problem for the inline JIT code generated as that uses the
shape's realm and ignores the context.

Differential Revision: https://phabricator.services.mozilla.com/D130643

UltraBlame original commit: 59dbf9d365e5e3fd78383630e897243bba14c6ba