Bug 1732343 - Part 9: Directly use DuplicateHandle in RemoteBackbuffer, r=cmartin
Previously this code was using ipc::DuplicateHandle, which was unnecessary, as
the code is being run in the parent process which has access to perform direct
handle transfers. After this patch we'd like to remove ipc::DuplicateHandle and
rely only on attached handles for serialization, so the code is converted to
instead directly use DuplicateHandle.

Differential Revision: https://phabricator.services.mozilla.com/D126571

UltraBlame original commit: 1d94da94e88807ee078bae5c32dd09dadd4c9716