Bug 1735976 - Update Firefox Suggest preferences: Part 2 - Update the preferences UI. r=Gijs
This updates the Privacy pane for the new Firefox Suggest preferences spec.

Please see the spec here: https://mozilla-hub.atlassian.net/browse/SNT-37

The Jira description is a little out of date, and the wording of these new
strings (which are not exposed to localizers) isn't finalized, but I'm told the
structure of the UI is final more or less.

There is also a Figma here: https://www.figma.com/file/seJ2ZA4v3FgoV7jCxUR74B/Firefox-Suggest-exploration?node-id=3197%3A55695

We're replacing the current two Firefox Suggest checkboxes with three toggle
buttons. The first two toggle buttons correspond to the existing
`browser.urlbar.suggest.quicksuggest` and
`browser.urlbar.suggest.quicksuggest.sponsored` prefs. However, the second pref
is no longer dependent on the first, and it can be toggled regardless of whether
the first is enabled. The third toggle corresponds to a new pref,
`browser.urlbar.quicksuggest.dataCollection.enabled`. It can also be toggled
independently of the others.

In addition, we're adding an info bar/box below the toggles to explain to the
user the effect of their toggle selection. The text in the box depends on the
state of the toggles. The box itself is hidden when all three toggles are off.

Depends on D129224

Differential Revision: https://phabricator.services.mozilla.com/D128661

UltraBlame original commit: d9093f96237d3ed715a4affcaeb6461be9d8451d